+++ tests/unit/Model/Command/UserForgetTest.php
@@ -2,6 +2,7 @@
 
 namespace Lmc\Matej\Model\Command;
 
+use Lmc\Matej\Model\Command\Constants\UserForgetMethod;
 use Lmc\Matej\UnitTestCase;
 
 class UserForgetTest extends UnitTestCase
@@ -12,10 +13,10 @@ class UserForgetTest extends UnitTestCase
         $userId = 'user-id';
 
         $command = UserForget::anonymize($userId);
-        $this->assertForgetCommand($command, $userId, UserForget::ANONYMIZE);
+        $this->assertForgetCommand($command, $userId, UserForgetMethod::ANONYMIZE());
 
         $command = UserForget::delete($userId);
-        $this->assertForgetCommand($command, $userId, UserForget::DELETE);
+        $this->assertForgetCommand($command, $userId, UserForgetMethod::DELETE());
     }
 
     /**
@@ -23,7 +24,7 @@ class UserForgetTest extends UnitTestCase
      *
      * @param UserForget $command
      */
-    private function assertForgetCommand($command, string $userId, string $method): void
+    private function assertForgetCommand($command, string $userId, UserForgetMethod $method): void
     {
         $this->assertInstanceOf(UserForget::class, $command);
         $this->assertSame(
@@ -31,12 +32,12 @@ class UserForgetTest extends UnitTestCase
                 'type' => 'user-forget',
                 'parameters' => [
                     'user_id' => $userId,
-                    'method' => $method,
+                    'method' => $method->jsonSerialize(),
                 ],
             ],
             $command->jsonSerialize()
         );
         $this->assertSame($userId, $command->getUserId());
-        $this->assertSame($method, $command->getForgetMethod());
+        $this->assertEquals($method, $command->getForgetMethod());
     }
 }