+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - **BC BREAK** | `UserRecommendation` now returns new format of response in `->getData()`, which is a list of `stdClass` instances.
 - **BC BREAK** | `UserRecommendation` does not have default filter (was previously set to: `valid_to >= NOW`).
 - **BC BREAK** | `UserRecommendation` now uses MQL query language by default for filtering.
+- **BC BREAK** | Minimal relevance passed to `setMinimalRelevance()` method of `UserRecommendation` command must be of enum type `MinimalRelevance`.
+- **BC BREAK** | `getForgetMethod()` method of `UserForget` command now returns instance of `UserForgetMethod` enum instead of string.
 
 ### Fixed
 - Exceptions occurring during async request now generate rejected promise (as they should) and are no longer thrown directly.